chore: move bokeh from top level file
